Tidy-up. No functionality changes
